# Add MessagePackReader with support for ReadOnlySequence<byte>