--- a/sandbox/DynamicCodeDumper/DynamicCodeDumper.csproj
+++ b/sandbox/DynamicCodeDumper/DynamicCodeDumper.csproj
@@ -58,8 +58,26 @@
     <Compile Include="..\..\src\MessagePack\MessagePackBinary.cs">
       <Link>Code\MessagePackBinary.cs</Link>
     </Compile>
+    <Compile Include="..\..\src\MessagePack\ExtensionHeader.cs">
+      <Link>Code\ExtensionHeader.cs</Link>
+    </Compile>
+    <Compile Include="..\..\src\MessagePack\ExtensionResult.cs">
+      <Link>Code\ExtensionResult.cs</Link>
+    </Compile>
+    <Compile Include="..\..\src\MessagePack\MessagePackReader.cs">
+      <Link>Code\MessagePackReader.cs</Link>
+    </Compile>
+    <Compile Include="..\..\src\MessagePack\SequenceReader.cs">
+      <Link>Code\SequenceReader.cs</Link>
+    </Compile>
+    <Compile Include="..\..\src\MessagePack\SequenceReaderExtensions.cs">
+      <Link>Code\SequenceReaderExtensions.cs</Link>
+    </Compile>
     <Compile Include="..\..\src\MessagePack\MessagePackCode.cs">
       <Link>Code\MessagePackCode.cs</Link>
+    </Compile>
+    <Compile Include="..\..\src\MessagePack\Internal\DateTimeConstants.cs">
+      <Link>Code\Internal\DateTimeConstants.cs</Link>
     </Compile>
     <Compile Include="..\..\src\MessagePack\Nil.cs">
       <Link>Code\Nil.cs</Link>
@@ -81,6 +99,8 @@
     </Compile>
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="System.Memory" Version="4.5.2" />
+    <PackageReference Include="Nerdbank.Streams" Version="2.0.214" />
+    <PackageReference Include="System.Memory" Version="4.5.1" />
+    <PackageReference Include="System.Runtime.CompilerServices.Unsafe" Version="4.5.2" />
   </ItemGroup>
 </Project>