--- a/sandbox/DynamicCodeDumper/Program.cs
+++ b/sandbox/DynamicCodeDumper/Program.cs
@@ -4,6 +4,7 @@ using MessagePack.Internal;
 using MessagePack.Resolvers;
 using SharedData;
 using System;
+using System.Buffers;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
@@ -125,9 +126,9 @@ namespace DynamicCodeDumper
     }
     public class Int_x10Formatter : IMessagePackFormatter<int>
     {
-        public int Deserialize(byte[] bytes, int offset, IFormatterResolver formatterResolver, out int readSize)
+        public int Deserialize(ref MessagePackReader reader, IFormatterResolver formatterResolver)
         {
-            return MessagePackBinary.ReadInt32(bytes, offset, out readSize) * 10;
+            return reader.ReadInt32() * 10;
         }
 
         public int Serialize(ref byte[] bytes, int offset, int value, IFormatterResolver formatterResolver)
@@ -138,9 +139,9 @@ namespace DynamicCodeDumper
 
     public class String_x2Formatter : IMessagePackFormatter<string>
     {
-        public string Deserialize(byte[] bytes, int offset, IFormatterResolver formatterResolver, out int readSize)
+        public string Deserialize(ref MessagePackReader reader, IFormatterResolver formatterResolver)
         {
-            var s = MessagePackBinary.ReadString(bytes, offset, out readSize);
+            var s = reader.ReadString();
             return s + s;
         }