--- a/sandbox/PerfBenchmarkDotNet/PerfBenchmarkDotNet.csproj
+++ b/sandbox/PerfBenchmarkDotNet/PerfBenchmarkDotNet.csproj
@@ -1,9 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <RootNamespace>PerfBenchmarkDotNet</RootNamespace>
-    <AssemblyName>PerfBenchmarkDotNet</AssemblyName>
-    <TargetFramework>net461</TargetFramework>
+    <TargetFrameworks>net47;netcoreapp2.0</TargetFrameworks>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
   </PropertyGroup>
   <ItemGroup>
@@ -14,25 +12,16 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="BenchmarkDotNet.Diagnostics.Windows" version="0.10.8" />
-    <PackageReference Include="BenchmarkDotNet.Toolchains.Roslyn" version="0.10.8" />
+    <PackageReference Include="BenchmarkDotNet" version="0.11.4" />
+    <PackageReference Include="BenchmarkDotNet.Diagnostics.Windows" Version="0.11.4" />
     <PackageReference Include="Hyperion" version="0.9.6" />
     <PackageReference Include="Jil" version="2.15.4" />
     <PackageReference Include="MessagePack" version="1.4.3" />
-    <PackageReference Include="Microsoft.CodeAnalysis.Analyzers" version="1.1.0" />
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" version="2.0.0" />
-    <PackageReference Include="Microsoft.DotNet.InternalAbstractions" version="1.0.0" />
-    <PackageReference Include="Microsoft.DotNet.PlatformAbstractions" version="1.1.1" />
-    <PackageReference Include="Microsoft.Win32.Registry" version="4.3.0" />
     <PackageReference Include="MsgPack.Cli" version="0.9.0-rc1" />
     <PackageReference Include="Newtonsoft.Json" version="10.0.3" />
     <PackageReference Include="protobuf-net" version="2.3.2" />
     <PackageReference Include="Sigil" version="4.7.0" />
     <PackageReference Include="ZeroFormatter" version="1.6.4" />
-  </ItemGroup>
-  <ItemGroup>
-    <Content Include="MessagePack_1_4_4.dll" />
-    <Content Include="MessagePack_1_6_0_3.dll" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\src\MessagePack.Annotations\MessagePack.Annotations.csproj">