--- a/src/MessagePack.ImmutableCollection/MessagePack.ImmutableCollection.csproj
+++ b/src/MessagePack.ImmutableCollection/MessagePack.ImmutableCollection.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>netstandard1.6;netstandard2.0;net45;net47</TargetFrameworks>
+    <TargetFrameworks>netstandard1.6;netstandard2.0;net46;net47</TargetFrameworks>
 
     <IsPackable>true</IsPackable>
     <Title>MessagePack for C# Extension Support for ImmutableCollection</Title>