… README.md
… sandbox/Sandbox/Generated.cs
… sandbox/Sandbox/Program.cs
… sandbox/SharedData/Class1.cs
… src/MessagePack.ReactiveProperty/Formatters.cs
… src/MessagePack.UnityShims/Extension/UnsafeBlitFormatter.cs
… src/MessagePack.UnityShims/Formatters.cs
… src/MessagePack.UniversalCodeGenerator/CodeAnalysis/Definitions.cs
… src/MessagePack.UniversalCodeGenerator/Generator/EnumTemplate.cs
… src/MessagePack.UniversalCodeGenerator/Generator/EnumTemplate.tt
… src/MessagePack.UniversalCodeGenerator/Generator/FormatterTemplate.cs
… src/MessagePack.UniversalCodeGenerator/Generator/FormatterTemplate.tt
… src/MessagePack.UniversalCodeGenerator/Generator/ResolverTemplate.cs
… src/MessagePack.UniversalCodeGenerator/Generator/ResolverTemplate.tt
… src/MessagePack.UniversalCodeGenerator/Generator/UnionTemplate.cs
… src/MessagePack.UniversalCodeGenerator/Generator/UnionTemplate.tt
… src/MessagePack/ExtensionHeader.cs
… src/MessagePack/ExtensionResult.cs
… src/MessagePack/Formatters/CollectionFormatter.cs
… src/MessagePack/Formatters/DictionaryFormatter.cs
… src/MessagePack/Formatters/DynamicObjectTypeFallbackFormatter.cs
… src/MessagePack/Formatters/EnumAsStringFormatter.cs
… src/MessagePack/Formatters/ForceSizePrimitiveFormatter.cs
… src/MessagePack/Formatters/ForceSizePrimitiveFormatter.tt
… src/MessagePack/Formatters/IMessagePackFormatter.cs
… src/MessagePack/Formatters/IgnoreFormatter.cs
… src/MessagePack/Formatters/MultiDimentionalArrayFormatter.cs
… src/MessagePack/Formatters/NullableFormatter.cs
… src/MessagePack/Formatters/OldSpecFormatter.cs
… src/MessagePack/Formatters/PrimitiveFormatter.cs