This replaces the (byte[]+offset+out bytesRead) reading methods.

These changes are a subset extracted from #27 

Toward overall resolution of #9