@AArnott 2019-02-24
@neuecc I plan to merge this soon so I can follow it up with a `MessagePackWriter` PR. Do you have time to review it in the next 36-48 hours?