@neuecc 2019-02-28
I'm sorry, I misunderstood you.
I thought that PR would come to neuecc/MessagePack-CSharp,
Now is the story to review this PR?