@neuecc 2019-02-28
I've started review in 2 areas, performance/api.

Performance:
I've uploaded latest msgpack(1_7_3_4) to PerfBenchmarkDotNet.
At first, run your ReadByte10 vs ReadByte20 test.

```
BenchmarkDotNet=v0.11.4, OS=Windows 10.0.16299.846 (1709/FallCreatorsUpdate/Redstone3)
Intel Core i9-7980XE CPU 2.60GHz, 1 CPU, 36 logical and 18 physical cores
Frequency=2531250 Hz, Resolution=395.0617 ns, Timer=TSC
  [Host] : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3260.0
  Clr    : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3260.0
  Core   : .NET Core 2.1.6 (CoreCLR 4.6.27019.06, CoreFX 4.6.27019.05), 64bit RyuJIT


|     Method |  Job | Runtime | Categories |     Mean |     Error |    StdDev |
|----------- |----- |-------- |----------- |---------:|----------:|----------:|
| ReadByte20 |  Clr |     Clr |        2.0 | 4.829 ns | 0.0920 ns | 0.0984 ns |
| ReadByte20 | Core |    Core |        2.0 | 4.093 ns | 0.0434 ns | 0.0406 ns |
|            |      |         |            |          |           |           |
| ReadByte10 |  Clr |     Clr |        1.0 | 3.273 ns | 0.0222 ns | 0.0197 ns |
| ReadByte10 | Core |    Core |        1.0 | 4.068 ns | 0.0245 ns | 0.0229 ns |
```

I'll add more tests and show result tomorrow.

API:
Review and comment tomorrow.