@AArnott 2019-02-28
> I thought that PR would come to neuecc/MessagePack-CSharp,

All of the v2.0 changes are still in my fork at the moment.  When you're OK with it, I'll branch off v1.8 in your repo from your current master branch, then start sending PRs from my master to your (new) master to get all my changes into your repo.

> Now is the story to review this PR?

Yes please. I'm taking this large change a piece at a time, and this PR is one of those pieces.