@AArnott 2019-02-28
Regarding the perf, on .NET Core the regression is within the margin of error if you consider those columns. I don't think we have anything to worry about there.

On .NET Framework, it's around 1.5ns slower to specifically read a byte. There *might* be something we can optimize there, but unless there's a legit scenario that reads lots of individual bytes, I don't think it'll add up and cause a problem for anyone. These are *nanoseconds* we're talking about here, after all. What do you think?