@neuecc 2019-03-01
API:
Should OldSpec compatibility do on Reader?
I thinks it should only care/select in formatter and resolver.
I think, should remove `MessagePackReader.OldSpec` property.

Is ReadBytes returns `ReadOnlySequence<byte>` safe for buffer pool?
Current(1.7.3.4) impl decide it is unsafe so split Readbytes and ReadSegment.
`ExtensionResult.Data` too.

`Skip` is `ReadNextBlock`, so where is `ReadNext` compatibility method?

ReadInt32(and others) can not read Int64 format.

```csharp
byte[] bytes = null;
oldmsgpack::MessagePack.MessagePackBinary.WriteInt64ForceInt64Block(ref bytes, 0, 1);
var reader = new newmsgpack::MessagePack.MessagePackReader(bytes);
Console.WriteLine(reader.ReadInt32()); // exception
```

it is bad, i think this should read and use `checked` cast.

---

Performance:

> I don't think we have anything to worry about there.

Yes, I think too.

But reading small single byte's perfomrance is too bad.

```csharp
[CategoriesColumn]
[Config(typeof(BenchmarkConfig))]
public class SingleMessagePackReaderBenchmark
{
    private byte[] buffer = new byte[1];

    [GlobalSetup]
    public void GlobalSetup()
    {
        this.buffer[0] = newmsgpack::MessagePack.MessagePackCode.MaxFixInt;
    }

    [Benchmark, BenchmarkCategory("2.0")]
    public int SingleReadInt20()
    {
        return new newmsgpack::MessagePack.MessagePackReader(this.buffer).ReadInt32();
    }

    [Benchmark, BenchmarkCategory("1.0")]
    public int SingleReadInt10()
    {
        var offset = 0;
        var result = oldmsgpack::MessagePack.MessagePackBinary.ReadInt32(this.buffer, offset, out int readSize);
        offset += readSize;
        return result;
    }
}
```

```
BenchmarkDotNet=v0.11.4, OS=Windows 10.0.16299.846 (1709/FallCreatorsUpdate/Redstone3)
Intel Core i9-7980XE CPU 2.60GHz, 1 CPU, 36 logical and 18 physical cores
Frequency=2531250 Hz, Resolution=395.0617 ns, Timer=TSC
  [Host] : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3260.0
  Clr    : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3260.0
  Core   : .NET Core 2.1.6 (CoreCLR 4.6.27019.06, CoreFX 4.6.27019.05), 64bit RyuJIT

Platform=X64  IterationCount=1  WarmupCount=1

|          Method |  Job | Runtime | Categories |       Mean | Error | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|---------------- |----- |-------- |----------- |-----------:|------:|------------:|------------:|------------:|--------------------:|
| SingleReadInt20 |  Clr |     Clr |        2.0 | 120.780 ns |    NA |           - |           - |           - |                   - |
| SingleReadInt10 |  Clr |     Clr |        1.0 |   3.028 ns |    NA |           - |           - |           - |                   - |
| SingleReadInt20 | Core |    Core |        2.0 | 102.682 ns |    NA |           - |           - |           - |                   - |
| SingleReadInt10 | Core |    Core |        1.0 |   3.703 ns |    NA |           - |           - |           - |                   - |
```

There are many scenarios to read small bytes.
(read from KVS(Redis, Memcached), read small argument parameter(in RPC), etc...).
If the result of this benchmark is correct, it is critical issue in 2.0.

also check from `Deserialize` path.

```csharp
[CategoriesColumn]
[Config(typeof(BenchmarkConfig))]
public class ReadFromDeserialize
{
    private byte[] buffer = new byte[1];
    newmsgpack::MessagePack.MessagePackSerializer serializer;

    [GlobalSetup]
    public void GlobalSetup()
    {
        this.buffer[0] = newmsgpack::MessagePack.MessagePackCode.MaxFixInt;
        this.serializer = new newmsgpack::MessagePack.MessagePackSerializer();
    }

    [Benchmark, BenchmarkCategory("2.0")]
    public int SingleReadInt20()
    {
        return serializer.Deserialize<int>(buffer);
    }

    [Benchmark, BenchmarkCategory("1.0")]
    public int SingleReadInt10()
    {
        return oldmsgpack::MessagePack.MessagePackSerializer.Deserialize<int>(buffer);
    }
}
```

```
BenchmarkDotNet=v0.11.4, OS=Windows 10.0.16299.846 (1709/FallCreatorsUpdate/Redstone3)
Intel Core i9-7980XE CPU 2.60GHz, 1 CPU, 36 logical and 18 physical cores
Frequency=2531250 Hz, Resolution=395.0617 ns, Timer=TSC
  [Host] : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3260.0
  Clr    : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3260.0
  Core   : .NET Core 2.1.6 (CoreCLR 4.6.27019.06, CoreFX 4.6.27019.05), 64bit RyuJIT

Platform=X64  IterationCount=1  WarmupCount=1

|          Method |  Job | Runtime | Categories |      Mean | Error | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|---------------- |----- |-------- |----------- |----------:|------:|------------:|------------:|------------:|--------------------:|
| SingleReadInt20 |  Clr |     Clr |        2.0 | 154.82 ns |    NA |           - |           - |           - |                   - |
| SingleReadInt10 |  Clr |     Clr |        1.0 |  20.79 ns |    NA |           - |           - |           - |                   - |
| SingleReadInt20 | Core |    Core |        2.0 | 114.61 ns |    NA |           - |           - |           - |                   - |
| SingleReadInt10 | Core |    Core |        1.0 |  17.41 ns |    NA |           - |           - |           - |                   - |
```